+++ eq-author/src/components/AnswerContent/AdvancedProperties/index.js
