+++ eq-author/src/components/AnswerContent/AnswerProperties/Answers/CheckboxProperties.js
