+++ eq-author/src/components/AnswerContent/AnswerProperties/Answers/DurationProperties.js
