+++ eq-author/src/components/AnswerContent/AnswerProperties/Answers/DurationProperties.test.js
