+++ eq-author/src/components/AnswerContent/AnswerProperties/Answers/NumberProperties.js
