+++ eq-author/src/components/AnswerContent/AnswerProperties/Decimal/index.test.js
